Redid units spawning in zones
Sort of added an alarm when caught inside a zone. It doesn't seem to be
loud enough and needs to be tweaked.
Tweaked the Zone QRF function so that it requires it to be fufilled for
5-10 seconds before firing, allowing players to possibly stealth their
way through a zone.
Missed a pair of brackets in unlocked uniforms.
Converted enemy spawns in zones at game start into functions. This seems
to massively increase performance.