Major update
So vehicle spawning was majorly overhauled. Now vehicles can be placed
much more precisely, due to having a "ghost" vehicle to use as a
preview. FOBs and the HQ can also be placed in this manner. There are
also two different vehicle spawn lists now, one that uses CP as normal,
and another one that uses USD, which is a currency which will be
expanded on soon. USD is implied to be our budget from the merc group we
work with.